requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
Flask==2.2.5
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
proto-plus==1.26.1
protobuf==4.24.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9.1
typing_extensions==4.7.1
urllib3==1.26.20
Werkzeug==2.2.3
zipp==3.15.0
